Tooltip placement considers edge of viewport
By default, a tooltip is placed to the right and below the mouse cursor.
This caused the tooltip to get cut off at the right and lower edge of
the viewport.

With this change, before the tooltip is rendered, we calculate if the
sum of max-width or max-height of the tooltip plus the current mouse position
would overshoot the edge of the viewport. If so, the tooltip is placed
on the opposite side of the mouse.